requirements.txt
matplotlib==3.6.3
numpy==1.23.5
opencv_contrib_python==4.7.0.68
Pillow==9.4.0
pyheif==0.7.1
